requirements.txt
bottle
pyyaml
paste
